Added bigdcd flag for pbc wrap
* (pbcwrap) pbcwral.tcl: added "-bigdcd" flag to make the pbc wrap
command work when used inside a proc to be used with bigdcd. The flag
is equivalent to "-now", but bypasses the animate command at the end of
pbc wrap.

The animate command at the end of pbc wrap wich is used to rewind to
the frame the wrap was started with is conflicting with the animate
commands used by bigdcd.tcl to load frames of a dcd one at a time into
vmd. The "-bigdcd" flag uses an if statement to bypass the animate
command and prints out and also info statement if "-verbose" used.

* doc/pbctools.tex: Added description of "-bigdcd" to documentation
explaining that it is equivalent to "-now" and intended for use inside
a proc that will be called with bigdcd.

* AUTHORS: Added myself to the authors. Remove if contribution not
substantial enough to merit addition.